condaenv.5psig9sb.requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
chembl-webresource-client==0.10.9
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
easydict==1.13
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
ipywidgets==8.1.5
jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-ollama==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.24
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
more-click==0.1.2
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
orjson==3.10.16
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pyparsing==3.2.3
pystow==0.7.0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
seaborn==0.13.2
sniffio==1.3.1
sqlalchemy==2.0.40
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.0
triton==3.2.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.3.0
widgetsnbextension==4.0.13
wordcloud==1.9.4
yarl==1.19.0
zstandard==0.23.0